requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pyyaml
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
openai
python-dotenv